update for change computer
update

+++ Portfolio-architecte-sophie-bluel-master/FrontEnd/assets/script.js
@@ -12,24 +12,25 @@ let categDataName;
 
 
 // on récupère les données de l'API Works
-
-fetch(apiWorks)
-    .then(reponse => {
-        if (!reponse.ok) {
-            throw new Error(`Erreur HTTP! statut: ${reponse.status}`);
-        }
-        return reponse.json();
-    })
-    .then(data => {
-        projets = data;
-        createButtons();
-        mesProjets();
-        modeEditionPhotos()       
-    })
-    .catch(error => {
-        console.error("Une erreur s'est produite lors de la récupération des données", error);
-    });
-
+const getworks = () => {
+    fetch(apiWorks)
+        .then(reponse => {
+            if (!reponse.ok) {
+                throw new Error(`Erreur HTTP! statut: ${reponse.status}`);
+            }
+            return reponse.json();
+        })
+        .then(data => {
+            projets = data;
+            createButtons();
+            mesProjets();
+            setModal()
+        })
+        .catch(error => {
+            console.error("Une erreur s'est produite lors de la récupération des données", error);
+        });
+}
+getworks()
 // on récupère les données de l'API Categories
 
 fetch(apiCategory)
@@ -73,8 +74,21 @@ function mesProjets() {
 const divFiltres = document.querySelector(".filtres");
 
 function createButtons() {
+    divFiltres.innerHTML = ""
     const uniqueCategoryIds = [];
 
+    const boutonTous = document.createElement("button")
+    divFiltres.appendChild(boutonTous)
+    boutonTous.classList.add("filtres-boutton")
+    boutonTous.innerText = "Tous"
+    boutonTous.addEventListener("click", function () {
+        category = 0;
+        mesProjets();
+        const boutonsFiltres = document.querySelectorAll(".filtres-boutton");
+        boutonsFiltres.forEach(bouton => bouton.classList.remove("filtres-boutton_selected"));
+        // Mettez à jour la classe du bouton "Tous"
+        boutonTous.classList.add("filtres-boutton_selected");
+    });
     for (let i = 0; i < projets.length; i++) {
         const categoryId = projets[i].categoryId;
         if (!uniqueCategoryIds.includes(categoryId)) {
@@ -82,7 +96,6 @@ function createButtons() {
         }
     }
 
-    const divFiltres = document.querySelector(".filtres");
 
     for (let index = 0; index < uniqueCategoryIds.length; index++) {
         const categoryId = uniqueCategoryIds[index];
@@ -93,24 +106,34 @@ function createButtons() {
         boutonFiltres.innerText = categoryData.category.name; // Utilisez le nom de la catégorie
 
         boutonFiltres.addEventListener("click", () => {
-            category = categoryId; // Utilisez l'ID de catégorie
-            mesProjets();
+            category = categoryId
+            mesProjets()
+            const boutonsFiltres = document.querySelectorAll(".filtres-boutton");
+            boutonsFiltres.forEach(bouton => bouton.classList.remove("filtres-boutton_selected"));
+            // Mettez à jour la classe du bouton actuel
+            boutonFiltres.classList.add("filtres-boutton_selected");
+
         });
 
         divFiltres.appendChild(boutonFiltres);
     }
 }
 
-//gestion du bouton tous 
-const boutonTous = document.getElementById("tous");
-boutonTous.addEventListener("click", function () {
-    category = 0;
-    mesProjets();
-});
+// Suite à la création des boutons, ont leur applique leur animation 
+
+
 
 // relie le DOM
+
 const introduction = document.getElementById("introduction");
 const portFolio = document.getElementById("portfolio");
+// on créée des boutons édition fictif 
+
+const btnEditIntro = document.createElement("button")
+introduction.appendChild(btnEditIntro)
+btnEditIntro.classList.add("boutonEditionMenu")
+btnEditIntro.innerText = "Modifier"
+
 
 //mode édition
 // afficher mode édition 
@@ -132,13 +155,13 @@ function initAddEventListenerPopup() {
     btnEdit = document.querySelector(".boutonEdition");
     btnEdit.addEventListener("click", () => {
         afficherPopup();
-        modeEditionPhotos();
+        setModal();
     });
     cross.addEventListener("click", () => {
         cacherPopup();
     });
     arrowLeft.addEventListener("click", () => {
-        modeEditionPhotos()
+        setModal()
         arrowLeft.classList.add("disabled")
     })
 
@@ -146,18 +169,17 @@ function initAddEventListenerPopup() {
 initAddEventListenerPopup();
 
 // on génère les photos dans le mode edit 
-const galleryEdit = document.querySelector(".popup-galerie");
+const ModalGallery = document.querySelector(".popup-galerie");
 const popupTitre = document.querySelector(".popup-titre");
 
-// on met en place le popup modeEditionPhotos
-function modeEditionPhotos() {
-    galleryEdit.innerHTML = "";
+// on met en place le popup setModal
+function setModal() {
+    ModalGallery.innerHTML = "";
 
     for (let i = 0; i < projets.length; i++) {
 
         const photos = projets[i]
-photos.id
-        const galleryEditPhotos = document.createElement("figure");
+        const ModalGalleryPhotos = document.createElement("figure");
         const galleryPhotos = document.createElement("img");
 
         const buttonDeletePhotos = document.createElement("button");
@@ -175,45 +197,55 @@ photos.id
         buttonDeletePhotos.addEventListener("click", () => {
             console.log(`${photos.id}`)
 
-            // API supprimer photo 
-            const info = JSON.parse(localStorage.getItem("info"))
-            if (info) {
-                const { userId, token } = info;
-
-                fetch(`http://localhost:5678/api/works/${photos.id}`, {
-                    method: "DELETE",
-                    headers: {
-                        "Authorization": `Bearer ${token}`,
-                        "Content-Type": "application/json"
-                    }
-                })
-                    .then(response => {
-                        if (response.ok) {
-                            console.log("Suppression réussie. ")
-                        } else {
-                            console.error("Echec de la suppresion")
-                        }
-                    })
-                    .catch(error => {
-                        console.error("Erreur lors de la suppresion:", error)
-                    })
-            } else {
-                console.error("Aucune information d\'autorisation trouvée dans le localStorage")
-            }
+            deletWork(photos.id)
 
         });
 
         galleryPhotos.src = photos.imageUrl;
 
-        galleryEdit.appendChild(galleryEditPhotos);
-        galleryEditPhotos.appendChild(galleryPhotos);
-        galleryEditPhotos.appendChild(buttonDeletePhotos);
+        ModalGallery.appendChild(ModalGalleryPhotos);
+        ModalGalleryPhotos.appendChild(galleryPhotos);
+        ModalGalleryPhotos.appendChild(buttonDeletePhotos);
     }
 
     popupTitre.innerText = "supprimer des photos";
     bouttonAddPhotos.classList.remove("disabled");
 }
 
+const deletWork = (id) => {
+
+    if (confirm("Êtes vous sûr de vouloir supprimer cette photo ?") == true) {
+
+
+        // API supprimer photo 
+        const info = JSON.parse(localStorage.getItem("info"))
+        if (info) {
+            const { userId, token } = info;
+
+            fetch(`http://localhost:5678/api/works/${id}`, {
+                method: "DELETE",
+                headers: {
+                    "Authorization": `Bearer ${token}`,
+                    "Content-Type": "application/json"
+                }
+            })
+                .then(response => {
+                    if (response.ok) {
+                        getworks()
+                        console.log("Suppression réussie. ")
+
+                    } else {
+                        console.error("Echec de la suppresion")
+                    }
+                })
+                .catch(error => {
+                    console.error("Erreur lors de la suppresion:", error)
+                })
+        } else {
+            console.error("Aucune information d\'autorisation trouvée dans le localStorage")
+        }
+    }
+}
 // on rajoute eventListener pour le clic sur le bouton 
 // on crée le bouton pour ajouter les photos 
 
@@ -225,13 +257,14 @@ bouttonAddPhotos.innerText = "Ajouter une photo";
 
 //on génére la page d'ajout des photos
 
-function ajouterPhotos(categData) {
-    galleryEdit.innerHTML = ""
+function ajouterPhotos() {
+
+    ModalGallery.innerHTML = ""
     popupTitre.innerText = "Ajouter une photo"
     arrowLeft.classList.remove("disabled")
 
     const popupForm = document.createElement("form")
-    galleryEdit.appendChild(popupForm)
+    ModalGallery.appendChild(popupForm)
 
     const labelImage = document.createElement("label");
     labelImage.setAttribute("for", "photoImage");
@@ -281,16 +314,11 @@ function ajouterPhotos(categData) {
 
     bouttonAddPhotosSubmit.addEventListener("click", (event) => {
         event.preventDefault();
-        // for (let i = 0 ; i < projetsId ; i++){
-        //     photoId = [i]
-        // }
-        // console.log(photoId)
-        // Vérification des champs obligatoires
         if (
             photoTitre.value.trim() === "" ||
             photoImage.files.length === 0 ||
             selectCategory.value === ""
-            
+
         ) {
             alert("Veuillez remplir tous les champs obligatoires.");
         } else {
@@ -300,41 +328,42 @@ function ajouterPhotos(categData) {
 
                 const formData = new FormData();
                 formData.append("title", photoTitre.value);
-                formData.append("imageUrl", photoImage.files[0]);
-                formData.append("categoryId", selectCategory.value);
-                formData.append("userId", userId);
+                formData.append("image", photoImage.files[0]);
+                formData.append("category", selectCategory.value);
 
                 fetch(`http://localhost:5678/api/works`, {
                     method: "POST",
                     body: formData,
                     headers: {
                         "Authorization": `Bearer ${token}`,
-                        "content-type":"application/json",
+                        "accept": "application/json",
                     },
                 })
-                .then(response => {
-                    if (!response.ok) {
-                        throw new Error('La requête a échoué.');
-                    }
-                    return response.json();
-                })
-                .then(data => {
-                    // "data" contient le contenu de la réponse JSON
-                    // Vous pouvez maintenant travailler avec les données
-                    photo.push(data); // Ajoutez la nouvelle photo à la liste
-                    localStorage.setItem("photos", JSON.stringify(data)); // Enregistrez la liste mise à jour
-                    console.log("succès")
+                    .then(response => {
+                        if (!response.ok) {
+                            throw new Error('La requête a échoué.');
+                        }
+                        return response.json();
+                    })
+                    .then(data => {
+                        // "data" contient le contenu de la réponse JSON
+                        // Vous pouvez maintenant travailler avec les données
+                        // projets.push(data); // Ajoutez la nouvelle photo à la liste
+                        // localStorage.setItem("photos", JSON.stringify(data)); // Enregistrez la liste mise à jour
+                        // console.log("succès")
 
-                    mesProjets();
-                    console.log("succès")
-                })
-                .catch(error => {
-                    console.error(error);
-                });
+                        // mesProjets();
+                        // console.log("succès")
+                        getworks()
+
+                    })
+                    .catch(error => {
+                        console.error(error);
+                    });
                 console.log(photoImage.files)
                 console.log(photoTitre.value)
                 console.log(selectCategory.value)
-               
+
             }
         }
     });
@@ -344,6 +373,10 @@ function ajouterPhotos(categData) {
 
 
 bouttonAddPhotos.addEventListener("click", () => {
-    ajouterPhotos(categData)
+    ajouterPhotos()
 })
+
+function previsuPhotos()  {
+
+}
